+++ src/mimic/src/schema/node/field.rs
@@ -1,7 +1,7 @@
 use crate::{
     schema::{
         build::validate::validate_ident,
-        node::{Arg, ValidateNode, Value, VisitableNode},
+        node::{Arg, TypeValidator, ValidateNode, Value, VisitableNode},
         visit::Visitor,
     },
     types::{Cardinality, ErrorVec, SortDirection},
@@ -64,8 +64,11 @@ pub struct Field {
     pub name: String,
     pub value: Value,
 
-    #[serde(skip_serializing_if = "Option::is_none")]
+    #[serde(default, skip_serializing_if = "Option::is_none")]
     pub default: Option<Arg>,
+
+    #[serde(default, skip_serializing_if = "Vec::is_empty")]
+    pub validators: Vec<TypeValidator>,
 }
 
 impl ValidateNode for Field {
@@ -111,6 +114,9 @@ impl VisitableNode for Field {
     fn drive<V: Visitor>(&self, v: &mut V) {
         self.value.accept(v);
         if let Some(node) = &self.default {
+            node.accept(v);
+        }
+        for node in &self.validators {
             node.accept(v);
         }
     }